Moved notification builders to their own package
