+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -80,10 +80,10 @@ import org.odk.collect.android.metadata.InstallIDProvider;
 import org.odk.collect.android.metadata.SharedPreferencesInstallIDProvider;
 import org.odk.collect.android.network.ConnectivityProvider;
 import org.odk.collect.android.network.NetworkStateProvider;
-import org.odk.collect.android.notifications.FormUpdatesAvailableNotificationBuilder;
-import org.odk.collect.android.notifications.FormUpdatesDownloadedNotificationBuilder;
-import org.odk.collect.android.notifications.FormsSubmissionNotificationBuilder;
-import org.odk.collect.android.notifications.FormsSyncFailedNotificationBuilder;
+import org.odk.collect.android.notifications.builders.FormUpdatesAvailableNotificationBuilder;
+import org.odk.collect.android.notifications.builders.FormUpdatesDownloadedNotificationBuilder;
+import org.odk.collect.android.notifications.builders.FormsSubmissionNotificationBuilder;
+import org.odk.collect.android.notifications.builders.FormsSyncFailedNotificationBuilder;
 import org.odk.collect.android.notifications.NotificationManagerNotifier;
 import org.odk.collect.android.notifications.Notifier;
 import org.odk.collect.android.openrosa.CollectThenSystemContentTypeMapper;
