+++ collect_app/src/main/java/org/odk/collect/android/notifications/NotificationManagerNotifier.kt
@@ -7,6 +7,10 @@ import android.content.Context
 import android.os.Build
 import org.odk.collect.android.R
 import org.odk.collect.android.formmanagement.ServerFormDetails
+import org.odk.collect.android.notifications.builders.FormUpdatesAvailableNotificationBuilder
+import org.odk.collect.android.notifications.builders.FormUpdatesDownloadedNotificationBuilder
+import org.odk.collect.android.notifications.builders.FormsSubmissionNotificationBuilder
+import org.odk.collect.android.notifications.builders.FormsSyncFailedNotificationBuilder
 import org.odk.collect.android.preferences.keys.MetaKeys
 import org.odk.collect.android.preferences.source.SettingsProvider
 import org.odk.collect.forms.FormSourceException
